admin_core/admin_console/requirements.txt
Flask==2.3.5
PyYAML==6.0.1
gunicorn==22.0.0
boto3==1.35.10
cryptography==43.0.1
requests==2.32.3
pyjwt==2.9.0
aws-secretsmanager-caching==1.1.3
flask-login==0.6.3
flask-caching==2.1.0
redis==5.0.8
pyotp==2.9.0
argon2-cffi==23.1.0
flask-talisman==1.1.0